+++ lib/groups/index.ts
@@ -3,3 +3,5 @@ export * from './permissions';
 export * from './schema';
 export * from './assignments';
 export * from './classifications';
+export * from './memberPerformance';
+export * from './aiMetadata';